cleaned up struc extract code
